feat(item): implement lead item and add `Leashable` trait (#355)
