+++ steel-core/src/entity/living_entity.rs
@@ -567,15 +567,6 @@ pub trait LivingEntity: Entity {
         world.clip(start, end, block_shape, fluid).is_miss()
     }
 
-    /// Returns vanilla base living-entity invulnerability.
-    fn default_is_invulnerable_to(&self, source: &DamageSource) -> bool {
-        self.is_removed()
-            || self.is_invulnerable() && !source.bypasses_invulnerability()
-            || source.is(&vanilla_damage_type_tags::DamageTypeTag::IS_FIRE) && self.fire_immune()
-            || source.is(&vanilla_damage_type_tags::DamageTypeTag::IS_FALL)
-                && self.is_fall_damage_immune()
-    }
-
     /// Returns whether this living entity ignores a damage source.
     fn is_invulnerable_to(&self, world: &World, source: &DamageSource) -> bool {
         self.default_is_invulnerable_to(source)
